requirements.txt
certifi==2026.2.25
charset-normalizer==3.4.4
colorzero==2.0
dotenv==0.9.9
gpiozero==2.0.1
idna==3.11
lgpio==0.2.2.0
pillow==12.1.1
python-dotenv==1.2.2
qrcode==8.2
requests==2.32.5
RPi.GPIO==0.7.1
spidev==3.8
urllib3==2.6.3
websockets==16.0
